+++ toolium/test_cases.py
@@ -135,12 +135,10 @@ class SeleniumTestCase(BasicTestCase):
         super(SeleniumTestCase, self).tearDown()
         test_name = self.get_subclassmethod_name().replace('.', '_')
 
-        # Capture screenshot on error
+        # Capture screenshot and save logs on error
         if not self._test_passed:
             DriverWrappersPool.capture_screenshots(test_name)
-
-        # Write Webdriver logs to files
-        self.utils.save_all_webdriver_logs(self.get_subclassmethod_name())
+            DriverWrappersPool.save_all_webdriver_logs(self.get_subclassmethod_name())
 
         # Close browser and stop driver if it must not be reused
         DriverWrappersPool.close_drivers_and_download_videos(test_name, self._test_passed, self.reuse_driver)